test(ethernetip): move EtherNet/IP tests into their own folder
Consolidate the EtherNet/IP test suite under tests/ethernetip/: the
test_ethernetip_* modules, the shared fakes, the wheel smoke script, the
cpppo simulator helper, and the typecheck fixtures. Update intra-package
imports to tests.ethernetip.*, fix the repo-root depth in the type-stub
test, and point the justfile recipes at the new paths.

INSTRO-411
