+++ tests/ethernetip/test_ethernetip_config_data_type.py
@@ -6,7 +6,7 @@ import pytest
 from pydantic import ValidationError
 
 from instro.ethernetip import EtherNetIPConfig, EtherNetIPDevice
-from tests.ethernetip_fakes import FakePlcKind, FakePlcValue, install_fake_native_ethernetip
+from tests.ethernetip.ethernetip_fakes import FakePlcKind, FakePlcValue, install_fake_native_ethernetip
 
 
 def test_data_type_is_required_for_tags() -> None:
