+++ tests/ethernetip/typecheck/ethernetip_valid.py
